requirements.txt
altgraph==0.17.4
amqp==5.2.0
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.2.10
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
billiard==4.2.1
boto3==1.35.25
botocore==1.35.25
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
DBUtils==3.1.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.1.2
django-bulk-update-or-create==0.3.0
django-extensions==3.1.5
djangorestframework==3.15.1
docopt==0.6.2
drf-yasg==1.21.7
dropbox==12.0.2
et-xmlfile==1.1.0
executing==2.1.0
fastkml==0.12
fonttools==4.54.1
future==1.0.0
geojson==3.1.0
gevent==24.2.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
ipython==8.27.0
ipython-genutils==0.2.0
isodate==0.6.1
itypes==1.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jdcal==1.4.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
keyring==25.4.1
kiwisolver==1.4.7
kombu==5.4.2
linecache2==1.0.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
mssql-django==1.2
numpy==2.1.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pefile==2024.8.26
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygeoif==0.7
pyglet==2.0.17
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
Pympler==1.1
PyMySQL==1.1.1
pyodbc==5.1.0
pyparsing==3.1.4
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
scipy==1.14.1
selenium==4.25.0
sendgrid==6.11.0
setuptools==75.1.0
six==1.16.0
smopy==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
stack-data==0.6.3
starkbank-ecdsa==2.2.0
stomp-py==8.1.2
stone==3.3.1
traceback2==1.4.0
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Unidecode==1.3.8
unittest2==1.1.0
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
Wand==0.6.13
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.44.0
whitenoise==6.7.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.20.2
zope.event==5.0
zope.interface==7.0.3